Colors RED (as Danger/Warning) was changed to dark green for winnig player.
Winning is often not braught into relation with RED. RED means Danger/Warning/Alert. So I changed winning Player to Green and loosing player to blue. at the same Time, I added constants for the colors. Better would be a simple Java-Color-Chooser Dialog and then store the colors in the Config-File of the user.
Shall I do that?